# Conflicts:
#	src/main/java/com/gofobao/framework/tender/service/impl/TenderServiceImpl.java

+++ src/main/java/com/gofobao/framework/comment/entity/Topic.java
@@ -26,7 +26,7 @@ public class Topic implements Serializable {
 
     private String title;
 
-    private Integer topicTypeId;
+    private long topicTypeId;
 
     private long userId;
